Combine country/provider/sector dashboards
